improve performance with rowRender memo
